Explicit casts for args arising from main()
kvlds/dispatch.c:
	The main binary checked that opt_k and opt_v (which turn into
	D->kmax and D->vmax) are between 0 and 255.

lbs-dynamodb/state.c:
	The main binary checked that opt_b (which turns into itemsz) is
	not more than 8192.  That said, I figured there's no harm
	ensuring that the cast is safe in case anybody uses state_init()
	elsewhere.

lbs-s3/s3state.c:
	The main binary already checked that opt_b (which becomes
	blklen) is between 512 and 128 * 1024 (or 2^17).  I added a
	sanity check as well, just in case.

+++ kvlds/dispatch.c
@@ -479,7 +479,7 @@ gotrequest(void * cookie, int status)
 		case PROTO_KVLDS_PARAMS:
 			/* Send the response immediately. */
 			if (proto_kvlds_response_params(D->writeq, RQ->R->ID,
-			    D->kmax, D->vmax))
+			    (uint32_t)D->kmax, (uint32_t)D->vmax))
 				goto err2;
 
 			/* Free the linked list node. */
